pyproject.toml
[tool.poetry]
name = 'tevmc'
version = '0.1a5'
description = 'Telos EVM Docker controller / framework'
authors = ['Guillermo Rodriguez <guillermo@telos.net>']
readme = 'README.md'

[tool.poetry.dependencies]
python = '^3.9'
web3 = '^6.8.0'
tdqm = '^0.0.1'
asks = '^3.0.0'
flask = '^2.3.3'
click = '^8.1.7'
pytest = '^7.4.0'
psutil = '^5.9.5'
docker = '^6.1.3'
urllib3 = '1.26.18'
natsort = '^8.4.0'
requests = '^2.31.0'
iterators = '^0.2.0'
simplejson = '^3.19.1'
elasticsearch = '^8.9.0'
requests-unixsocket = '^0.3.0'
py-leap = {git = 'https://github.com/guilledk/py-leap', rev = 'v0.1a17'}
rlp = '3.0.0'
pdbplus = "^1.5.0"

[build-system]
requires = ['poetry-core']
build-backend = 'poetry.core.masonry.api'

[tool.poetry.scripts]
tevmc = 'tevmc.cmdline:cli'

[build]
include = ['tevmc/templates/**/*']

[tool.pytest.ini_options]
log_cli = true
log_cli_level = 'INFO'
markers = [
    'config: tevmc.json config for test',
    'randomize: enables/disables port & creds randomization',
    'services: which services should run',
    'custom_nodeos_tar: install a custom nodeos binary from a tar',
    'tevmc_params: pass parameters to TEVMController constructor'
]
